requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
google-genai
jsonpatch
pytest
httpx
python-multipart
boto3
mangum
sqlalchemy
pyodbc
